Moved all store-related files to a folder with an index and updated all imports
