+++ ui/src/lib/store/index.ts
@@ -1,0 +1,4 @@
+export type { GamezClient, BoardDefData } from './store';
+export { GamezStore } from './store';
+export { type Piece, PieceDef, PieceType, type BoardProps, type BoardState, Board } from './board';
+export { BoardType } from './boardList';
